requirements.txt
Flask
Flask-Cors
mysql-connector-python
qrcode
pillow
flask-bcrypt
PyJWT
python-dotenv